Fix issue when checking guest IP of the VM with docker running
When docker daemon runs in the VM, there will be a nic created by
docker and usually has IP like 172.x.x.x. When VM is booting up,
VM Tools sometimes reports this IP to vCenter, although the real IP
of the non-docker nic will be reported later.
So we add a check for this case and filter out the docker IP.